Manual cleanup of quote marks in ase.db.sqlite
It is discouraged to use double-quotes inside sqlite statements, and
officially unsupported (by default) for recent versions.

Some users/developers have reported errors related to the use of
double-quote marks in sqlite statements, but they are proving
difficult to reproduce in a clean environment.

In the ASE codebase it is standard to quote strings with
single-quotes.

In the context of this file, which was mostly quoted with
double-quotes, the ASE convention should be to use single-quotes
_unless_ the string contains a SQL statement with a quoted string: in
such cases the python string should use double quotes and the SQL
statement use single-quotes internally. With a consistent style in the
rest of the file, this helps these strings to stand out slightly as
having an interesting quote property.